時間の型をdoubleに

+++ app/src/main/java/com/sasakulab/yure_android_client/MainActivity.kt
@@ -154,7 +154,7 @@ data class AccelerationData(
     val x: Double,
     val y: Double,
     val z: Double,
-    val t: Long,
+    val t: Double,
     val userAgent: String,
 )
